Split sortable stories into two js modules
Seperate grid and resizer stories.
